requirements.txt
streamlit
google-gemini
